git-svn-id: http://svn.osgeo.org/mapguide/trunk@9318 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/BuildTools/WebTools/IMake/SimpleXmlParser.cpp
@@ -180,7 +180,7 @@ const char* _tcsenistr( const char* psz, const char* str, int len, int escape )
 
 const char* _tcseistr( const char* psz, const char* str, int escape )
 {
-    int len = strlen( str );
+    size_t len = strlen( str );
     return _tcsenistr( psz, str, len, escape );
 }
